csrf protection to journaltheme delete related #609

+++ src/Ojs/JournalBundle/Controller/JournalThemeController.php
@@ -9,6 +9,7 @@ use Symfony\Component\HttpFoundation\Request;
 use Ojs\Common\Controller\OjsController as Controller;
 use Ojs\JournalBundle\Entity\JournalTheme;
 use Ojs\JournalBundle\Form\JournalThemeType;
+use Symfony\Component\Security\Core\Exception\TokenNotFoundException;
 
 /**
  * JournalTheme controller.
@@ -27,6 +28,8 @@ class JournalThemeController extends Controller
         $grid = $this->get('grid')->setSource($source);
 
         $actionColumn = new ActionsColumn("actions", 'actions');
+        ActionHelper::setup($this->get('security.csrf.token_manager'));
+
         $rowAction[] = ActionHelper::showAction('admin_journaltheme_show', 'id');
         $rowAction[] = ActionHelper::editAction('admin_journaltheme_edit', 'id');
         $rowAction[] = ActionHelper::deleteAction('admin_journaltheme_delete', 'id');
@@ -187,14 +190,20 @@ class JournalThemeController extends Controller
     }
 
     /**
-     * Deletes a JournalTheme entity.
-     * @param  JournalTheme                                       $entity
+     * @param Request $request
+     * @param JournalTheme $entity
      * @return \Symfony\Component\HttpFoundation\RedirectResponse
+     * @throws TokenNotFoundException
      */
-    public function deleteAction(JournalTheme $entity)
+    public function deleteAction(Request $request, JournalTheme $entity)
     {
         $this->throw404IfNotFound($entity);
         $em = $this->getDoctrine()->getManager();
+
+        $csrf = $this->get('security.csrf.token_manager');
+        $token = $csrf->getToken('admin_journaltheme'.$entity->getId());
+        if($token!=$request->get('_token'))
+            throw new TokenNotFoundException("Token Not Found!");
         $em->remove($entity);
         $em->flush();
         $this->successFlashBag('successful.remove');
